requirements.txt
python-dotenv
requests
feedparser
newspaper3k
lxml_html_clean
anthropic
